testes/foca-backend/requirements.txt
fastapi
motor
python-dotenv
pydantic
uvicorn
